test: update unit test routes + mock jwt to use in all unit test routes
